turned carrousel into server component
